backend/src/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
phonenumbers==8.13.14
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
requests==2.31.0
geopy==2.3.0
rollbar==0.16.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
